Improved Workflow and Schema check for JSON files
